pyproject.toml
[project]
name = "scrapping-ecommerce"
version = "0.1.0"
description = ""
authors = [
    {name = "William",email = "williamsilva14@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "selenium (>=4.31.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
